Added a function to find a device which starts with a given name
Added a function to find a device which starts with a given name.

The Xilinx Vivado and SDK creates a device tree file which contains the base address as a part of the name (e.g. "axi_bram_ctrl@6fff0000").
It is not possible to implement a general search for the base adresse of a device with "three byname(const char *name)" when the base address have to be part of the search string.